refactor(cloudwatch): share matcher helpers between alarm policies
Extract AlarmMatcher / AlarmMatchContext / matchesOne / ruleMatches into
policies/policy-matcher.ts so the alarm-name and alarm-actions policies
both source the matcher contract from a single place. Drops the existing
cross-policy import that leaked alarm-actions internals into alarm-name,
and eliminates ~25 lines of copy-paste.

Also stop publicly exporting the kebab() helper — it was only documented
as an implementation detail of defaultAlarmName.
