+++ packages/cloudwatch/src/index.ts
@@ -1,7 +1,7 @@
 export type { AlarmConfig, AlarmConfigDefaults } from "./alarm-config.js";
 export type { AlarmDefinition } from "./alarm-definition.js";
 export { AlarmDefinitionBuilder } from "./alarm-definition-builder.js";
-export { type AlarmName, alarmName, joinAlarmName, kebab } from "./alarm-name.js";
+export { type AlarmName, alarmName, joinAlarmName } from "./alarm-name.js";
 export { defaultAlarmName } from "./default-alarm-name.js";
 export { createAlarms } from "./create-alarms.js";
 export { resolveAlarmConfig, type ResolvedAlarmConfig } from "./resolve-alarm-config.js";
